+++ gui/shared.py
@@ -1,13 +1,3 @@
-import kivy
-
-kivy.require('2.0.0')
-
-from kivy.uix.popup import Popup
-from kivy.uix.label import Label
-from kivy.uix.button import Button
-from kivy.uix.boxlayout import BoxLayout
-
-
 class Shared:
 
     def __init__(self):
@@ -23,20 +13,3 @@ class Shared:
         self.app = None
         self.sm = None
         self.nm = None
-
-    @staticmethod
-    def popup_widget(popup_label, popup_title, close_button_label='Dismiss'):
-        layout = BoxLayout(orientation='vertical')
-        popup_content = Label(text=f'{popup_label}',
-                              valign='center', halign='center')
-        close_button = Button(text=f'{close_button_label}',
-                              size_hint=(None, None), size=(80, 40),
-                              pos_hint={'center_x': 0.5, 'center_y': 0.5})
-        layout.add_widget(popup_content)
-        layout.add_widget(close_button)
-        popup = Popup(title=f'{popup_title}', title_align='center',
-                      content=layout,
-                      size_hint=(None, None), size=(300, 200),
-                      auto_dismiss=False)
-        close_button.bind(on_press=popup.dismiss)
-        popup.open()
